Added UI, some changes
Added enum for facing direction.
Used this enum throught files.
Changed logic for interact (the ray starts on the back of player, which means that if the player stands on top of collectible, it will still be picked up.
Added UI text handling in new script, text displays which parts has been colected.